requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
mysql-connector-python
pymysql    
python-dotenv
#-e . 